backend/requirements.txt
Flask==2.3.2
PyJWT==2.7.0
Werkzeug==2.3.7
flask-cors==3.0.10
react-icons
typeface-roboto
ransformers
torch